[CP-XXX] Hook simplified by moving logic to analytics.ts
